requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
-e file:.#egg=cp2k-basis
    # via -r requirements.in
autopep8==2.0.0
    # via -r requirements.in
certifi==2022.9.24
    # via requests
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via flask
commonmark==0.9.1
    # via rich
deprecated==1.2.13
    # via limits
flake8==5.0.4
    # via
    #   -r requirements.in
    #   flake8-quotes
flake8-quotes==3.3.1
    # via -r requirements.in
flask==2.2.2
    # via
    #   cp2k-basis
    #   flask-limiter
flask-limiter==2.8.1
    # via cp2k-basis
h5py==3.7.0
    # via cp2k-basis
idna==3.4
    # via requests
importlib-metadata==5.0.0
    # via flask
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
limits==2.7.1
    # via flask-limiter
markupsafe==2.1.1
    # via
    #   jinja2
    #   werkzeug
mccabe==0.7.0
    # via flake8
numpy==1.23.4
    # via
    #   cp2k-basis
    #   h5py
packaging==21.3
    # via limits
pycodestyle==2.9.1
    # via
    #   autopep8
    #   flake8
pyflakes==2.5.0
    # via flake8
pygments==2.13.0
    # via rich
pyparsing==3.0.9
    # via packaging
pyyaml==6.0
    # via cp2k-basis
requests==2.28.1
    # via cp2k-basis
rich==12.6.0
    # via flask-limiter
tomli==2.0.1
    # via autopep8
typing-extensions==4.4.0
    # via
    #   flask-limiter
    #   limits
urllib3==1.26.12
    # via requests
werkzeug==2.2.2
    # via flask
wrapt==1.14.1
    # via deprecated
zipp==3.10.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
